+++ src/BoatTrackerBot/BoatTrackerBot.csproj
@@ -168,6 +168,7 @@
     </Compile>
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Utils\BookedSchedulerCache.cs" />
+    <Compile Include="Utils\Helpers.cs" />
   </ItemGroup>
   <ItemGroup>
     <Content Include="packages.config" />
